Moved util tests to common to match function locations
