+++ pkg/channel/distributed/common/config/secretwatcher.go
@@ -33,7 +33,7 @@ import (
 type SecretObserver func(ctx context.Context, secret *corev1.Secret)
 
 //
-// Initialize The Specified Context With A Secret Informer
+// InitializeSecretWatcher Initializes The Specified Context With A Secret Informer
 //
 func InitializeSecretWatcher(ctx context.Context, namespace string, name string, resyncTime time.Duration, observer SecretObserver) error {
