+++ pkg/channel/distributed/common/config/secretwatcher_test.go
@@ -24,12 +24,10 @@ import (
 
 	commonconstants "knative.dev/eventing-kafka/pkg/common/constants"
 
-	"github.com/Shopify/sarama"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
-	"knative.dev/eventing-kafka/pkg/common/config"
 	commontesting "knative.dev/eventing-kafka/pkg/common/testing"
 	injectionclient "knative.dev/pkg/client/injection/kube/client"
 	"knative.dev/pkg/system"
@@ -41,81 +39,12 @@ var (
 	secretMutex   = sync.Mutex{} // Don't trip up the data race examiner during tests
 )
 
-func TestGetAuthConfigFromKubernetes(t *testing.T) {
-	// Setup Test Environment Namespaces
-	commontesting.SetTestEnvironment(t)
-
-	// Test Data
-	oldAuthSecret := getSaramaTestSecret(t,
-		commontesting.SecretName,
-		commontesting.OldAuthUsername,
-		commontesting.OldAuthPassword,
-		commontesting.OldAuthNamespace,
-		commontesting.OldAuthSaslType)
-
-	// Define The TestCase Struct
-	type TestCase struct {
-		name    string
-		secret  *corev1.Secret
-		askName string
-		wantErr bool
-	}
-
-	// Create The TestCases
-	testCases := []TestCase{
-		{
-			name:    "Valid secret, valid request",
-			secret:  oldAuthSecret,
-			askName: commontesting.SecretName,
-			wantErr: false,
-		},
-		{
-			name:    "Valid secret, not found",
-			secret:  oldAuthSecret,
-			askName: "invalid-secret-name",
-			wantErr: true,
-		},
-	}
-
-	// Run The TestCases
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			ctx := context.WithValue(context.TODO(), injectionclient.Key{}, fake.NewSimpleClientset(testCase.secret))
-			kafkaAuth, err := config.GetAuthConfigFromKubernetes(ctx, testCase.askName, commontesting.SystemNamespace)
-			if !testCase.wantErr {
-				assert.Nil(t, err)
-				assert.NotNil(t, kafkaAuth)
-				assert.Equal(t, commontesting.OldAuthUsername, kafkaAuth.SASL.User)
-				assert.Equal(t, commontesting.OldAuthPassword, kafkaAuth.SASL.Password)
-				assert.Equal(t, commontesting.OldAuthSaslType, kafkaAuth.SASL.SaslType)
-			} else {
-				assert.NotNil(t, err)
-			}
-		})
-	}
-}
-
-func TestGetConfigFromSecret_Valid(t *testing.T) {
-	secret := getSaramaTestSecret(t,
-		commontesting.SecretName,
-		commontesting.OldAuthUsername,
-		commontesting.OldAuthPassword,
-		commontesting.OldAuthNamespace,
-		"")
-	kafkaAuth := config.GetAuthConfigFromSecret(secret)
-	assert.Equal(t, commontesting.OldAuthUsername, kafkaAuth.SASL.User)
-	assert.Equal(t, commontesting.OldAuthPassword, kafkaAuth.SASL.Password)
-	assert.Equal(t, sarama.SASLTypePlaintext, kafkaAuth.SASL.SaslType)
-}
-
-func TestGetConfigFromSecret_Invalid(t *testing.T) {
-	authCfg := config.GetAuthConfigFromSecret(nil)
-	assert.Nil(t, authCfg)
-}
-
 // Test The InitializeSecretWatcher() Functionality
 func TestInitializeSecretWatcher(t *testing.T) {
-	secret := getSaramaTestSecret(t,
+
+	// Add a Secret to a fake Kubernetes Client
+	commontesting.SetTestEnvironment(t)
+	secret := commontesting.GetTestSaramaSecret(
 		commontesting.SecretName,
 		commontesting.OldAuthUsername,
 		commontesting.OldAuthPassword,
@@ -181,10 +110,4 @@ func setWatchedSecret(secret *corev1.Secret) {
 func secretWatcherHandler(_ context.Context, secret *corev1.Secret) {
 	// Set the package variable to indicate that the test watcher was called
 	setWatchedSecret(secret)
-}
-
-func getSaramaTestSecret(t *testing.T, name string,
-	username string, password string, namespace string, saslType string) *corev1.Secret {
-	commontesting.SetTestEnvironment(t)
-	return commontesting.GetTestSaramaSecret(name, username, password, namespace, saslType)
 }
